requirements.txt
# Core MLflow for prompt registry and experiment tracking
mlflow>=3.0.0

# OpenAI API client
openai>=1.0.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Environment variables (.env file support)
python-dotenv>=1.0.0

# Progress bars for evaluation loops
tqdm>=4.66.0

# For automated prompt optimization (optional but recommended)
dspy-ai>=2.4.0

seaborn

plotly